refactor finished, all specs pass, and tested in the browser
