+++ app/services/c100_app/screener_decision_tree.rb
@@ -4,12 +4,26 @@ module C100App
       return next_step if next_step
 
       case step_name
-      # TODO: Put decision logic here
-      when :name
-        root_path
+      when :children_postcodes
+        check_if_court_is_valid
       else
         raise InvalidStep, "Invalid step '#{as || step_params}'"
       end
+    end
+
+    private
+
+    def check_if_court_is_valid
+      courts = CourtPostcodeChecker.new.courts_for(step_params.fetch(:children_postcodes))
+      if courts.empty?
+        show(:no_court_found)
+      else
+        edit('/steps/miam/consent_order')
+      end
+
+    # CourtPostcodeChecker already logs the exception
+    rescue StandardError
+      show(:error_but_continue)
     end
   end
 end
