+++ app/views/entrypoint/how_long.en.html.erb
@@ -20,7 +20,7 @@
       </div>
 
       <div class="xform-group form-submit">
-        <%= link_to 'Continue', edit_steps_postcode_screen_children_postcodes_path, class: 'button button-start', role: 'button' %>
+        <%= link_to 'Continue', edit_steps_screener_postcode_path, class: 'button button-start', role: 'button' %>
       </div>
     </div>
   </div>
