+++ config/locales/en.yml
@@ -639,8 +639,8 @@ en:
           heading: Can you provide evidence that your application is urgent?
           lead_text: You will need to provide evidence to support your claim
           or: or
-    postcode_screen:
-      children_postcodes:
+    screener:
+      postcode:
         edit:
           page_title: Where your children live
           heading: Where do your children live?
